perbaikan tampilan daftar usulan

+++ database/migrations/2025_12_02_155118_create_usulan_penilaian_table.php
@@ -10,21 +10,24 @@ return new class extends Migration
      * Run the migrations.
      */
     public function up(): void
-{
-    Schema::create('usulan_penilaian', function (Blueprint $table) {
-        $table->id();
+    {
+        // Hindari error jika tabel sudah ada
+        if (!Schema::hasTable('usulan_penilaian')) {
 
-        $table->foreignId('usulan_id')->constrained('usulans')->onDelete('cascade');
-        $table->foreignId('reviewer_id')->constrained('users')->onDelete('cascade');
-        $table->foreignId('komponen_id')->constrained('master_penilaian')->onDelete('cascade');
+            Schema::create('usulan_penilaian', function (Blueprint $table) {
+                $table->id();
+                
+                $table->foreignId('usulan_id')->constrained('usulans')->onDelete('cascade');
+                $table->foreignId('reviewer_id')->constrained('users')->onDelete('cascade');
+                $table->foreignId('komponen_id')->constrained('master_penilaian')->onDelete('cascade');
 
-        $table->integer('nilai')->default(0);
-        $table->text('catatan')->nullable();
+                $table->integer('nilai')->default(0);
+                $table->text('catatan')->nullable();
 
-        $table->timestamps();
-    });
-}
-
+                $table->timestamps();
+            });
+        }
+    }
 
     /**
      * Reverse the migrations.
